Refactored modules in SampleTagging. * Removed undeleted 'RecentScans.js' from the previous commit. * Moved Analysis.js, Search.js, and Tagging.js out from Analysis sub-folder.
