+++ src/screens/SampleTagging/Tagging.js
